docker-airflow-dags-redis-postgres/airflow/requirements.txt
python-dotenv>=0.21.0
oracledb>=1.3.0